+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- hnefatafl-client: don't unwrap() on sound errors..
+- hnefatafl-client: don't unwrap() on sound errors.
 
 ## [0.8.1] - 2025-03-12